+++ tencentcloud/cam/v20190116/models.go
@@ -5675,7 +5675,7 @@ type SubAccountUser struct {
 	// 最近登录IP
 	LastLoginIp *string `json:"LastLoginIp,omitnil" name:"LastLoginIp"`
 
-	// 最近登录时间
+	// 最近登录时间，回参为空，即为未登录过控制台
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	LastLoginTime *string `json:"LastLoginTime,omitnil" name:"LastLoginTime"`
 }