+++ tencentcloud/tke/v20180525/models.go
@@ -18225,11 +18225,14 @@ type ReservedInstanceSpec struct {
 }
 
 type ResourceDeleteOption struct {
-	// 资源类型，例如CBS
+	// 资源类型，例如CBS、CLB、CVM
 	ResourceType *string `json:"ResourceType,omitnil" name:"ResourceType"`
 
-	// 集群删除时资源的删除模式：terminate（销毁），retain （保留）
+	// 集群删除时CBS资源的删除模式：terminate（销毁），retain （保留）。其他资源默认为销毁。
 	DeleteMode *string `json:"DeleteMode,omitnil" name:"DeleteMode"`
+
+	// 是否跳过开启删除保护的资源，默认false，设置为true时不清理开启了删除保护的资源，clb有终端节点的情况也属于开了删除保护。
+	SkipDeletionProtection *bool `json:"SkipDeletionProtection,omitnil" name:"SkipDeletionProtection"`
 }
 
 type ResourceUsage struct {