+++ tencentcloud/vod/v20180717/models.go
@@ -23565,6 +23565,7 @@ type StatDataItem struct {
 	// <li>直播剪辑数据，单位是秒。</li>
 	// <li>轮播数据，单位是秒。</li>
 	// <li>子应用数量数据，单位是个。</li>
+	// <li>日志服务用量数据，单位是条。</li>
 	Value *int64 `json:"Value,omitnil" name:"Value"`
 }